added clinicservice

+++ SymptomCheck-backend/clinicservice/src/main/java/com/symptomcheck/clinicservice/repositories/MedicalClinicRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicalClinicRepository extends JpaRepository<MedicalClinic, Integer> {
+public interface MedicalClinicRepository extends JpaRepository<MedicalClinic, Long> {
 }
